Add files via upload
Distilbert and LSTM are two deep learning models that can be used to analyze large datasets for identifying and analyzing tweets containing hateful or derogatory speech. These models can also be used to detect hate speech in any kind of text sequence, not just limited to tweets. In this project, we attempted to use the concatenation of two datasets to evaluate the performance of Distilbert and LSTM in detecting such anomalies. Our results showed a high validation accuracy rate of 93% for Distilbert and 94% for LSTM, demonstrating the effectiveness of both models in detecting hate speech in tweets and other types of text sequences.